Organize client-api-schema types
I don't think everything is quite in the right place, but it is probably close enough.

My goal is to end up with individual method files that each export four types:
- `Params`
- `Request`
- `Response`
- `Errors` (a union type)
