+++ packages/client-api-schema/src/methods/ChallengeChannelRequest.ts
@@ -1,0 +1,6 @@
+import {JsonRpcRequest, JsonRpcResponse} from '../utils';
+import {ChannelId, ChannelResult} from '../data-types';
+
+export type ChallengeChannelParams = ChallengeChannelRequest['params']; // for backwards compatibility
+export type ChallengeChannelRequest = JsonRpcRequest<'ChallengeChannel', {channelId: ChannelId}>;
+export type ChallengeChannelResponse = JsonRpcResponse<ChannelResult>;
