+++ packages/client-api-schema/src/methods/GetBudgetRequest.ts
@@ -1,0 +1,6 @@
+import {JsonRpcRequest, JsonRpcResponse} from '../utils';
+import {Address, DomainBudget} from '../data-types';
+
+export type GetBudgetParams = {hubAddress: string};
+export type GetBudgetRequest = JsonRpcRequest<'GetBudget', {hubAddress: Address}>;
+export type GetBudgetResponse = JsonRpcResponse<DomainBudget | {}>;
