+++ packages/client-api-schema/src/methods/GetChannelsRequest.ts
@@ -1,0 +1,5 @@
+import {JsonRpcRequest, JsonRpcResponse} from '../utils';
+import {ChannelResult} from '../data-types';
+
+export type GetChannelsRequest = JsonRpcRequest<'GetChannels', {includeClosed?: boolean}>;
+export type GetChannelsResponse = JsonRpcResponse<ChannelResult[]>;
