+++ packages/client-api-schema/src/methods/GetWalletInformationRequest.ts
@@ -1,0 +1,9 @@
+import {JsonRpcRequest, JsonRpcResponse} from '../utils';
+import {Address} from '../data-types';
+
+export type GetWalletInformationRequest = JsonRpcRequest<'GetWalletInformation', {}>;
+export type GetWalletInformationResponse = JsonRpcResponse<{
+  signingAddress: Address;
+  selectedAddress: Address | null;
+  walletVersion: string;
+}>;
