+++ packages/client-api-schema/src/methods/PushMessageRequest.ts
@@ -1,0 +1,8 @@
+import {JsonRpcRequest, JsonRpcResponse} from '../utils';
+import {Message} from '../data-types';
+
+export type PushMessageParams = PushMessageRequest['params']; // included for backwards compatibility
+
+export type PushMessageResult = {success: boolean};
+export type PushMessageRequest = JsonRpcRequest<'PushMessage', Message>;
+export type PushMessageResponse = JsonRpcResponse<PushMessageResult>;
