+++ packages/client-api-schema/src/types.ts
@@ -1,258 +1,43 @@
-/**
- * Ethereum Address
- * @pattern  ^0x([a-fA-F0-9]{40})|0$
- */
-export type Address = string;
+import {
+  GetWalletInformationRequest,
+  GetWalletInformationResponse
+} from './methods/GetWalletInformationRequest';
 
-/**
- * Nitro ChannelId
- * @pattern  ^0x([a-fA-F0-9]{64})$
- */
-export type ChannelId = string;
+import {
+  EnableEthereumRequest,
+  EnableEthereumResponse,
+  EnableEthereumError
+} from './methods/EnableEthereumRequest';
 
-/**
- * Nitro ChannelId
- * @pattern  ^0x(0{24})([a-fA-F0-9]{40})$
- */
-export type ExternalDestination = string; // currently unused in this schema
+import {CreateChannelRequest, CreateChannelResponse} from './methods/CreateChannelRequest';
+import {JoinChannelRequest, JoinChannelResponse} from './methods/JoinChannelRequest';
+import {UpdateChannelRequest, UpdateChannelResponse} from './methods/UpdateChannelRequest';
+import {GetStateRequest, GetStateResponse} from './methods/GetStateRequest';
+import {PushMessageRequest, PushMessageResponse} from './methods/PushMessageRequest';
 
-/**
- * Bytes32
- * @pattern  ^0x([a-fA-F0-9]{64})$
- */
-export type Bytes32 = string;
+import {
+  CloseChannelRequest,
+  CloseChannelResponse,
+  NotYourTurnError
+} from './methods/CloseChannelRequest';
 
-/**
- * Uint256
- * @pattern  ^0x([a-fA-F0-9]{64})$
- */
-export type Uint256 = string;
+import {ChallengeChannelRequest, ChallengeChannelResponse} from './methods/ChallengeChannelRequest';
 
-interface JsonRpcRequest<MethodName, RequestParams> {
-  id: number; // in the json-rpc spec this is optional, but we require it for all our requests
-  jsonrpc: '2.0';
-  method: MethodName;
-  params: RequestParams;
-}
-interface JsonRpcResponse<ResultType> {
-  id: number;
-  jsonrpc: '2.0';
-  result: ResultType;
-}
+import {GetChannelsRequest, GetChannelsResponse} from './methods/GetChannelsRequest';
 
-interface JsonRpcNotification<NotificationName, NotificationParams> {
-  jsonrpc: '2.0';
-  method: NotificationName;
-  params: NotificationParams;
-}
+import {GetBudgetRequest, GetBudgetResponse} from './methods/GetBudgetRequest';
 
-export const UserDeclinedErrorCode = 200;
-export const EthereumNotEnabledErrorCode = 100;
-export const NotYourTurnErrorCode = 300;
+import {
+  ApproveBudgetAndFundRequest,
+  ApproveBudgetAndFundResponse
+} from './methods/ApproveBudgetAndFundRequest';
+import {
+  CloseAndWithdrawRequest,
+  CloseAndWithdrawResponse,
+  UserDeclinedErrorResponse
+} from './methods/CloseAndWithdrawParams';
 
-type ErrorCode =
-  | typeof UserDeclinedErrorCode
-  | typeof EthereumNotEnabledErrorCode
-  | typeof NotYourTurnErrorCode;
-interface JsonRpcError<Code extends ErrorCode, Message, Data = undefined> {
-  id: number;
-  jsonrpc: '2.0';
-  error: Data extends undefined
-    ? {code: Code; message: Message}
-    : {code: Code; message: Message; data: Data};
-}
-
-export type ChannelStatus =
-  | 'proposed'
-  | 'opening'
-  | 'funding'
-  | 'running'
-  | 'challenging'
-  | 'responding'
-  | 'closing'
-  | 'closed';
-
-export interface Participant {
-  participantId: string; // App allocated id, used for relaying messages to the participant
-  signingAddress: Address; // Address used to sign channel updates
-  destination: Address; // Address of EOA to receive channel proceeds (the account that'll get the funds).
-}
-
-export interface AllocationItem {
-  destination: Address; // Address of EOA to receive channel proceeds.
-  amount: Uint256; // How much funds will be transferred to the destination address.
-}
-
-export interface Allocation {
-  token: Address; // The token's contract address.
-  allocationItems: AllocationItem[]; // A list of allocations (how much funds will each destination address get).
-}
-
-export type Allocations = Allocation[]; // included for backwards compatibility
-
-export interface Message {
-  recipient: string; // Identifier of user that the message should be relayed to
-  sender: string; // Identifier of user that the message is from
-  data: unknown; // Message payload. Format defined by wallet and opaque to app.
-}
-
-export interface Funds {
-  token: Address;
-  amount: Uint256;
-}
-
-export interface ChannelResult {
-  participants: Participant[];
-  allocations: Allocation[];
-  appData: string;
-  appDefinition: Address;
-  channelId: ChannelId;
-  status: ChannelStatus;
-  turnNum: Uint256;
-  challengeExpirationTime?: number;
-}
-
-// GetWalletInformation
-export type GetWalletInformationRequest = JsonRpcRequest<'GetWalletInformation', {}>;
-export type GetWalletInformationResponse = JsonRpcResponse<{
-  signingAddress: Address;
-  selectedAddress: Address | null;
-  walletVersion: string;
-}>;
-
-// EnableEthereum
-export type EnableEthereumRequest = JsonRpcRequest<'EnableEthereum', {}>;
-export type EnableEthereumResponse = JsonRpcResponse<{
-  signingAddress: Address;
-  selectedAddress: Address;
-  walletVersion: string;
-}>;
-export type EnableEthereumError = JsonRpcError<
-  typeof EthereumNotEnabledErrorCode,
-  'Ethereum Not Enabled'
->;
-export type NotYourTurnError = JsonRpcError<typeof NotYourTurnErrorCode, 'Not your turn'>;
-
-// CreateChannel
-export type FundingStrategy = 'Direct' | 'Ledger' | 'Virtual';
-export interface CreateChannelParams {
-  participants: Participant[];
-  allocations: Allocation[];
-  appDefinition: Address;
-  appData: string;
-  fundingStrategy: FundingStrategy;
-}
-export type CreateChannelRequest = JsonRpcRequest<'CreateChannel', CreateChannelParams>;
-export type CreateChannelResponse = JsonRpcResponse<ChannelResult>;
-
-// JoinChannel
-export interface JoinChannelParams {
-  channelId: ChannelId;
-}
-export type JoinChannelRequest = JsonRpcRequest<'JoinChannel', JoinChannelParams>;
-export type JoinChannelResponse = JsonRpcResponse<ChannelResult>;
-
-// UpdateChannel
-export interface UpdateChannelParams {
-  channelId: ChannelId;
-  participants: Participant[];
-  allocations: Allocation[];
-  appData: string;
-}
-export type UpdateChannelRequest = JsonRpcRequest<'UpdateChannel', UpdateChannelParams>;
-export type UpdateChannelResponse = JsonRpcResponse<ChannelResult>;
-
-// GetState
-export interface GetStateParams {
-  channelId: ChannelId;
-}
-export type GetStateRequest = JsonRpcRequest<'GetState', GetStateParams>;
-export type GetStateResponse = JsonRpcResponse<ChannelResult>;
-
-// PushMessage
-export type PushMessageParams = PushMessageRequest['params']; // included for backwards compatibility
-export type PushMessageResult = {success: boolean};
-export type PushMessageRequest = JsonRpcRequest<'PushMessage', Message>;
-export type PushMessageResponse = JsonRpcResponse<PushMessageResult>;
-
-// CloseChannel
-export interface CloseChannelParams {
-  channelId: ChannelId;
-}
-export type CloseChannelRequest = JsonRpcRequest<'CloseChannel', CloseChannelParams>;
-export type CloseChannelResponse = JsonRpcResponse<ChannelResult>;
-
-// ChallengeChannel
-export type ChallengeChannelParams = ChallengeChannelRequest['params']; // for backwards compatibility
-export type ChallengeChannelRequest = JsonRpcRequest<'ChallengeChannel', {channelId: ChannelId}>;
-export type ChallengeChannelResponse = JsonRpcResponse<ChannelResult>;
-
-// Budget
-
-export interface ChannelBudget {
-  channelId: Bytes32;
-  amount: Uint256;
-}
-
-export interface TokenBudget {
-  token: Address;
-  availableReceiveCapacity: Uint256;
-  availableSendCapacity: Uint256;
-  channels: ChannelBudget[];
-}
-export interface DomainBudget {
-  domain: string;
-  hubAddress: string;
-  budgets: TokenBudget[];
-}
-
-export interface TokenBudgetRequest {
-  hub: Participant;
-  playerParticipantId: string;
-  token: Address;
-  requestedSendCapacity: Uint256;
-  requestedReceiveCapacity: Uint256;
-}
-
-export type GetChannelsRequest = JsonRpcRequest<'GetChannels', {includeClosed?: boolean}>;
-export type GetChannelsResponse = JsonRpcResponse<ChannelResult[]>;
-export type GetBudgetRequest = JsonRpcRequest<'GetBudget', {hubAddress: Address}>;
-export type GetBudgetResponse = JsonRpcResponse<DomainBudget | {}>;
-
-export type ApproveBudgetAndFundRequest = JsonRpcRequest<
-  'ApproveBudgetAndFund',
-  TokenBudgetRequest
->;
-export type ApproveBudgetAndFundResponse = JsonRpcResponse<DomainBudget>;
-
-export type CloseAndWithdrawParams = {hub: Participant; playerParticipantId: string};
-export type CloseAndWithdrawRequest = JsonRpcRequest<'CloseAndWithdraw', CloseAndWithdrawParams>;
-export type CloseAndWithdrawResponse = JsonRpcResponse<{success: boolean}>;
-
-export type GetBudgetParams = {hubAddress: string};
-// Notifications
-// these notifications come *from* the wallet, which is not strictly how JSON-RPC should work
-// (since we treat the wallet as the 'server')
-export type ChannelProposedNotification = JsonRpcNotification<'ChannelProposed', ChannelResult>;
-export type ChannelUpdatedNotification = JsonRpcNotification<'ChannelUpdated', ChannelResult>;
-export type ChannelClosingNotification = JsonRpcNotification<'ChannelClosed', ChannelResult>;
-export type MessageQueuedNotification = JsonRpcNotification<'MessageQueued', Message>;
-export type BudgetUpdatedNotification = JsonRpcNotification<'BudgetUpdated', DomainBudget>;
-export type UiNotification = JsonRpcNotification<'UIUpdate', {showWallet: boolean}>;
-
-export type Notification =
-  | ChannelProposedNotification
-  | ChannelUpdatedNotification
-  | ChannelClosingNotification
-  | BudgetUpdatedNotification
-  | MessageQueuedNotification
-  | UiNotification;
-
-type FilterByMethod<T, Method> = T extends {method: Method} ? T : never;
-
-export type NotificationType = {
-  [T in Notification['method']]: [FilterByMethod<Notification, T>['params']];
-};
+import {Notification} from './notifications';
 
 export type Request =
   | CreateChannelRequest
@@ -284,8 +69,6 @@ export type Response =
   | CloseAndWithdrawResponse
   | GetChannelsResponse;
 
-export type UserDeclinedErrorResponse = JsonRpcError<typeof UserDeclinedErrorCode, 'User declined'>;
-
 export type ErrorResponse = EnableEthereumError | UserDeclinedErrorResponse | NotYourTurnError;
 
 export type JsonRpcMessage = Request | Response | Notification | ErrorResponse;
@@ -304,3 +87,9 @@ export function isRequest(message: JsonRpcMessage): message is Request {
 export function isError(message: JsonRpcMessage): message is ErrorResponse {
   return 'id' in message && 'error' in message;
 }
+
+export * from './notifications';
+export * from './methods';
+export * from './data-types';
+
+export {EthereumNotEnabledErrorCode, UserDeclinedErrorCode} from './utils';
